@@ -5,7 +5,7 @@
     "minimum-stability": "stable",
     "prefer-stable": true,
     "require": {
-        "php": "^7.2.0",
+        "php": "^7.4",
         "ext-mbstring": "*",
         "codexsoft/cli": "^1.0",
         "codexsoft/code-classes": "^1.0",
